+++ plugins/css-audit/CHANGELOG.md
@@ -1,4 +1,7 @@
 v2.1.0
+- Audit now scans the compiler.options.output.path
+
+v2.1.0
 - Added error handling for if files aren't passed
 - Removed alphas from audits as it doesn't provide any results