+++ plugins/jshint/CHANGELOG.md
@@ -1,3 +1,7 @@
+v2.1.1
+- Audit now scans the compiler.options.output.path
+- Reporter updated to use @caweb/webpack/lib/handlebars
+
 v2.1.0
 - Plugin name was changed
 - Updated npm packages